Moved Admin API into its own servlet module
This permits deploying the Admin API without the OAuth2 API.
