+++ kangaroo-common/src/main/resources/logging.properties
