--- kangaroo-server/src/main/resources/hibernate.cfg.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2016 Michael Krotscheck
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
-  ~ use this file except in compliance with the License. You may obtain a copy
-  ~ of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<!DOCTYPE hibernate-configuration SYSTEM
-    "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-  <session-factory>
-    <!-- Connection properties -->
-    <property name="connection.datasource">java:comp/env/jdbc/OIDServerDB
-    </property>
-
-    <!-- Generic hibernate configuration -->
-    <property name="current_session_context_class">
-      org.hibernate.context.internal.ThreadLocalSessionContext
-    </property>
-    <property name="show_sql">false</property>
-    <property name="use_sql_comments">false</property>
-    <property name="generate_statistics">false</property>
-    <property name="order_updates">true</property>
-    <property name="max_fetch_depth">1</property>
-    <property name="default_batch_fetch_size">8</property>
-
-    <!-- Hibernate Search Configuration -->
-    <property name="search.default.directory_provider">filesystem</property>
-    <property name="hibernate.search.default.indexBase">/tmp/oid_lucene</property>
-    <property name="search.default.optimizer.operation_limit">1000</property>
-    <property name="search.default.optimizer.transaction_limit">100</property>
-    <property name="search.lucene_version">LUCENE_CURRENT</property>
-
-    <!-- Hibernate ID Gen Configuration -->
-    <property name="id.new_generator_mappings">true</property>
-
-    <!-- Entities -->
-    <mapping class="net.krotscheck.kangaroo.database.entity.Application"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.ApplicationScope"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.Authenticator"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.AuthenticatorState"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.Client"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.OAuthToken"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.Role"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.User"/>
-    <mapping class="net.krotscheck.kangaroo.database.entity.UserIdentity"/>
-
-  </session-factory>
-</hibernate-configuration>
