+++ kangaroo-server/src/main/webapp/WEB-INF/web.xml
@@ -22,7 +22,7 @@
     <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
     <init-param>
       <param-name>javax.ws.rs.Application</param-name>
-      <param-value>net.krotscheck.api.admin.v1.AdminV1API</param-value>
+      <param-value>net.krotscheck.kangaroo.servlet.admin.v1.AdminV1API</param-value>
     </init-param>
     <load-on-startup>1</load-on-startup>
   </servlet>
