+++ kangaroo-servlet-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/user/resource/package-info.java
@@ -15,29 +15,7 @@
  * limitations under the License.
  */
 
-package net.krotscheck.api.admin.v1.user;
-
-import net.krotscheck.api.admin.v1.user.resource.UserService;
-
-import javax.ws.rs.core.Feature;
-import javax.ws.rs.core.FeatureContext;
-
 /**
- * The User Management API.
- *
- * @author Michael Krotscheck
+ * Unit tests for the user service.
  */
-public final class UserFeature implements Feature {
-
-    /**
-     * Register this feature.
-     */
-    @Override
-    public boolean configure(final FeatureContext context) {
-
-        // User service
-        context.register(UserService.class);
-
-        return true;
-    }
-}
+package net.krotscheck.kangaroo.servlet.admin.v1.user.resource;
