+++ kangaroo-servlet-admin/src/test/resources/net/krotscheck/api/admin/v1/user/UserServiceTest.xml
