+++ pom.xml
@@ -19,6 +19,7 @@
     <module>kangaroo-authenticator-core</module>
     <module>kangaroo-authenticator-password</module>
     <module>kangaroo-servlet-oauth2</module>
+    <module>kangaroo-servlet-admin</module>
     <module>kangaroo-server</module>
   </modules>
